Moved tab fetching to scan module
